refactor: extract drawing utils and fix effect deps
